backend/requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
